Refactor node source parsing in its own class
